Split OpenSSL and non-OpenSSL parts of Public Key Crypto (RSA)
- pkenc_private_key_common.cpp split from pkenc_private_key.cpp
- pkenc_public_key_common.cpp  split from pkenc_public_key.cpp

This will allow other implementations, such as MBed TLS, in the future.

pkenc_*_key.cpp
- Set private_key_ and public_key_ to nullptr after freeing to prevent double frees
- Add #ifndef CRYPTOLIB_OPENSSL checks for OpenSSL-only code
- Fix incorrect comments and error messages

pktest.cpp
- Add test for BEGIN PUBLIC KEY header (instead of BEGIN RSA PUBLIC KEY header)
