+++ common/cpp/tests/Makefile
@@ -73,9 +73,11 @@ build/b64test: build build/b64test.o build/base64.o build/utils.o
 build/certtest: build build/certtest.o build/verify_certificate.o
 	g++ -o $@ $@.o           build/verify_certificate.o $(LDFLAGS)
 
-build/pktest: build build/pktest.o build/pkenc_private_key.o \
-		build/pkenc_public_key.o
-	g++ -o $@ $@.o build/pkenc_private_key.o build/pkenc_public_key.o \
+build/pktest: build build/pktest.o \
+		build/pkenc_private_key.o build/pkenc_private_key_common.o \
+		build/pkenc_public_key.o build/pkenc_public_key_common.o
+	g++ -o $@ $@.o build/pkenc_private_key.o build/pkenc_private_key_common.o \
+		build/pkenc_public_key.o build/pkenc_public_key_common.o \
  		$(LDFLAGS)
 
 build/signtest: build build/signtest.o \
